feat (wallet): synchronous auto spend and coin selection + max spend (#89)
* cleanup of some of the WalletSpend component

* add isSpendAll to test if tx might be spending all inputs

* disconnects `TransactionPreview` from redux store

* add getSpendableSlices selector

* update the naive coin selection utility
This makes coin selection somewhat more robust by determining if a change output is needed while still accounting for fees. It also makes use of spendAll utility to avoid unnecessary checks and adds tests for new functionality. Also includes some cleanup in components that will be using this.

* refactor(wallet): fix coin selection flow for auto spend
includes a lot of cleanup in the transaction reducer and action files. Also required some fixes to
the clearing of state when switching between views.

* feat(wallet): add max spend balance from the wallet in auto mode