feat(wallet): adds a check for status of next deposit slice after spend
This is meant to be a small usability enhancement for when someone does a self-spend to the next
unused deposit address, by checking and updating the status of that slice along with spent utxos and
change slice after a spend has been made. This will result in the UI updating to reflect this change
(both in the pending balance and in the slice table).