fix(wallet): update spending slices when WalletSign component unmounts (#98)
* fix(wallet): update spending slices when WalletSign component unmounts

this helps avoid an issue where two spends happen quickly back to back but the state of the
available utxos hasn't updated properly after spending from the first attempt. Waiting a little
gives the target, queried node time to process the previously broadcast tx and associated UTXOs

* feat(wallet): adds a check for status of next deposit slice after spend

This is meant to be a small usability enhancement for when someone does a self-spend to the next unused deposit address, by checking and updating the status of that slice along with spent utxos and change slice after a spend has been made. This will result in the UI updating to reflect this change (both in the pending balance and in the slice table).