feat(wallet): allow signing method to be 'unknown', and client empty
If a wallet config is uploaded that does not have a signing method included, caravan will set the
method to 'unknown' and during signing, the method selector dialog will be shown, but the BIP32 path
from the config will be used.

If a client isn't in a config, the ClientPicker will be shown.

BREAKING CHANGE:
none