fix(other): replace corsproxy with nginx reverse proxy (#210)
* fix(other): replace corsproxy with nginx reverse proxy

corsproxy is no longer maintained. A reference configuration for the nginx reverse proxy is provided
and installation instructions revised.

ISSUES: 209

* style(other): sanitize using generic username

Removed a specific username

* fix(other): fix CORS header implementation and revise nginx implementation

There was an error with the bitcoind.proxy file which prevented the CORS headers from being added

to the packets. The `nginx` implementation instructions were revised to increase clarity and add

the deprected `corsproxy` method.

ISSUES: 209

* docs(other): generalize nginx install instructions and fix minor errors

1. The installation instructions were generalized to other linux distributions with steps included
to identify
   the needed changes to `nginx.conf` to properly configure `nginx`.
2. There were some
typographical errors in the bash commands which were corrected.
3. Example test results were
included for the curl tests of the proxied/unproxied cases.
4. Clarified the username to be used is
the one specified in `rpcauth=`.

ISSUES: 209

* docs(other): fixed typos

Corrected the text `my_uname` to match the bash command in the `curl` examples.

ISSUES: 209